Refactor protocols modules
