+++ marlowe-chain-sync/libchainsync/Language/Marlowe/Runtime/ChainSync.hs
@@ -19,7 +19,7 @@ import Language.Marlowe.Runtime.ChainSync.Database (DatabaseQueries(..))
 import Language.Marlowe.Runtime.ChainSync.JobServer (ChainSyncJobServerDependencies(..), chainSyncJobServer)
 import Language.Marlowe.Runtime.ChainSync.QueryServer (ChainSyncQueryServerDependencies(..), chainSyncQueryServer)
 import Language.Marlowe.Runtime.ChainSync.Server (ChainSyncServerDependencies(..), chainSyncServer)
-import Network.Protocol.Driver (SomeConnectionSource)
+import Network.Protocol.Connection (SomeConnectionSource)
 import Network.Protocol.Job.Server (JobServer)
 import Network.Protocol.Query.Server (QueryServer)
 import Ouroboros.Network.Protocol.LocalTxSubmission.Client (SubmitResult)
