+++ marlowe-protocols/marlowe-protocols.cabal
@@ -55,6 +55,7 @@ library
     Network.Protocol.ChainSeek.Types
     Network.Protocol.Codec
     Network.Protocol.Codec.Spec
+    Network.Protocol.Connection
     Network.Protocol.Driver
     Network.Protocol.Handshake.Client
     Network.Protocol.Handshake.Server
@@ -66,6 +67,7 @@ library
     Network.Protocol.Query.Client
     Network.Protocol.Query.Server
     Network.Protocol.Query.Types
+    Observe.Event.Network.Protocol
   build-depends:
       base >= 4.9 && < 5
     , aeson
