+++ marlowe-protocols/src/Network/Protocol/Job/Types.hs
@@ -27,10 +27,10 @@ import Data.Type.Equality (type (:~:)(Refl))
 import GHC.Show (showSpace)
 import Network.Protocol.Codec (BinaryMessage(..))
 import Network.Protocol.Codec.Spec (ArbitraryMessage(..), MessageEq(..), ShowProtocol(..))
-import Network.Protocol.Driver (MessageToJSON(..))
 import Network.Protocol.Handshake.Types (HasSignature(..))
 import Network.TypedProtocol
 import Network.TypedProtocol.Codec (AnyMessageAndAgency(..))
+import Observe.Event.Network.Protocol (MessageToJSON(..))
 import Test.QuickCheck (Gen, oneof)
 
 data SomeTag cmd = forall status err result. SomeTag (Tag cmd status err result)
