+++ marlowe-protocols/src/Observe/Event/Network/Protocol.hs
@@ -1,0 +1,11 @@
+{-# LANGUAGE KindSignatures #-}
+{-# LANGUAGE PolyKinds #-}
+
+module Observe.Event.Network.Protocol
+  where
+
+import Data.Aeson (Value)
+import Network.TypedProtocol (Message, PeerHasAgency)
+
+class MessageToJSON ps where
+  messageToJSON :: PeerHasAgency pr (st :: ps) -> Message ps st st' -> Value
