+++ marlowe-runtime/cli/Language/Marlowe/Runtime/CLI/Command.hs
@@ -28,7 +28,8 @@ import Language.Marlowe.Runtime.CLI.Env (Env(..))
 import Language.Marlowe.Runtime.CLI.Monad (CLI, runCLI)
 import Language.Marlowe.Runtime.CLI.Option (optParserWithEnvDefault)
 import qualified Language.Marlowe.Runtime.CLI.Option as O
-import Network.Protocol.Driver (SomeConnector(..), tcpClient)
+import Network.Protocol.Connection (SomeConnector(..))
+import Network.Protocol.Driver (tcpClient)
 import Network.Protocol.Handshake.Client (handshakeClientConnector)
 import Network.Protocol.Job.Client (jobClientPeer)
 import Network.Socket (HostName, PortNumber)
