+++ marlowe-runtime/cli/Language/Marlowe/Runtime/CLI/Env.hs
@@ -8,7 +8,7 @@ module Language.Marlowe.Runtime.CLI.Env
 import Control.Concurrent.STM (STM)
 import Language.Marlowe.Protocol.Sync.Client (MarloweSyncClient)
 import Language.Marlowe.Runtime.Transaction.Api (MarloweTxCommand)
-import Network.Protocol.Driver (SomeClientConnector)
+import Network.Protocol.Connection (SomeClientConnector)
 import Network.Protocol.Job.Client (JobClient)
 
 -- | The environment for the Marlowe Runtime CLI.
