+++ marlowe-runtime/marlowe-indexer/Main.hs
@@ -24,7 +24,8 @@ import Language.Marlowe.Runtime.Indexer.Database (hoistDatabaseQueries)
 import qualified Language.Marlowe.Runtime.Indexer.Database.PostgreSQL as PostgreSQL
 import Logging (RootSelector(..), getRootSelectorConfig)
 import Network.Protocol.ChainSeek.Client (chainSeekClientPeer)
-import Network.Protocol.Driver (SomeConnector(..), logConnector, runConnector, tcpClient)
+import Network.Protocol.Connection (SomeConnector(..), logConnector)
+import Network.Protocol.Driver (runConnector, tcpClient)
 import Network.Protocol.Handshake.Client (handshakeClientConnector)
 import Network.Protocol.Query.Client (liftQuery, queryClientPeer)
 import Network.Socket (AddrInfo(..), HostName, PortNumber, SocketType(..), defaultHints, withSocketsDo)
