+++ marlowe-runtime/marlowe-sync/Main.hs
@@ -19,7 +19,8 @@ import Language.Marlowe.Runtime.Sync (SyncDependencies(..), sync)
 import Language.Marlowe.Runtime.Sync.Database (hoistDatabaseQueries, logDatabaseQueries)
 import qualified Language.Marlowe.Runtime.Sync.Database.PostgreSQL as Postgres
 import Logging (RootSelector(..), getRootSelectorConfig)
-import Network.Protocol.Driver (SomeConnectionSource(..), TcpServerDependencies(..), logConnectionSource, tcpServer)
+import Network.Protocol.Connection (SomeConnectionSource(..), logConnectionSource)
+import Network.Protocol.Driver (TcpServerDependencies(..), tcpServer)
 import Network.Protocol.Handshake.Server (handshakeConnectionSource)
 import Network.Socket (HostName, PortNumber)
 import Observe.Event.Backend (narrowEventBackend, newOnceFlagMVar)
