+++ marlowe-runtime/web-server/Language/Marlowe/Runtime/Web/Server.hs
@@ -36,7 +36,7 @@ import Language.Marlowe.Runtime.Web.Server.REST (ApiSelector)
 import qualified Language.Marlowe.Runtime.Web.Server.REST as REST
 import Language.Marlowe.Runtime.Web.Server.SyncClient (SyncClient(..), SyncClientDependencies(..), syncClient)
 import Language.Marlowe.Runtime.Web.Server.TxClient (TxClient(..), TxClientDependencies(..), TxClientSelector, txClient)
-import Network.Protocol.Driver (SomeClientConnector)
+import Network.Protocol.Connection (SomeClientConnector)
 import Network.Protocol.Job.Client (JobClient)
 import qualified Network.Wai as WAI
 import Network.Wai.Middleware.Cors (CorsResourcePolicy(..), cors, simpleCorsResourcePolicy)
